+++ src/features/usage-consolidation-accordion/usage-consolidation-content-package.js
@@ -11,7 +11,6 @@ import {
   Icon,
 } from '@folio/stripes/components';
 
-import NoCostPerUseAvailable from './no-cost-per-use-available';
 import SummaryTable from './summary-table';
 import { DEFAULT_SUMMARY_TABLE_COLUMNS } from './summary-table/column-properties';
 import TitlesTable from './titles-table';
@@ -164,12 +163,7 @@ const UsageConsolidationContentPackage = ({
     },
   };
 
-  return noCostPerUseAvailable ? (
-    <NoCostPerUseAvailable
-      entityType={entityTypes.PACKAGE}
-      year={year}
-    />
-  ) : (
+  return (
     <>
       <SummaryTable
         id="packageUsageConsolidationSummary"
@@ -181,6 +175,8 @@ const UsageConsolidationContentPackage = ({
         onViewTitles={onViewTitles}
         isExportDisabled={isExportDisabled}
         metricType={metricType}
+        noCostPerUseAvailable={noCostPerUseAvailable}
+        year={year}
       />
       <TitlesTable
         costPerUseData={costPerUseData}