+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 * Support erm interface v4.0
 * Fix typos in Full text requests table title. (UIEH-1039)
 * Title record: Usage & analysis > Summary Table > Remove Usage and CostPerUse columns. (UIEH-1040)
+* Fix message for empty cost and usage data in Resource Usage Consolidation Accordion. (UIEH-1041)
 
 ## [5.0.0] (https://github.com/folio-org/ui-eholdings/tree/v5.0.0) (2020-10-15)